Template: Load components dynamically and with overrides
Register all components dynamically at startup to the global Vue
object. First look for .vue files in a overrides directory (if it
exists). Then in the default components directory, skipping the ones
already loaded with same name.

After this there is no need to import components in views or other
components. But all components will be packed by webpack. So we must
be careful to use them all.

https://phabricator.endlessm.com/T31415