Template, scripts: Ensure overrides symlink
Remove symlink from git revision, and prevent its revision by adding
it to the ignored git files. Add script to ensure the symlink exists,
if not set it to the default override.

Ensure the symlink when running serve or build on the template
workspace.

Add a _common.py for shared constants and functions between scripts

Start documenting the scripts. The new one accepts `--help` for
usage.

https://phabricator.endlessm.com/T31415