Template: Paginate prop in CardGrid
Adds a paginate property to the CardGrid component to choose between
paginated or the previous version with a show more button. The
pagination can be configured also with the itemsPerPage property.

This patch also moves some code to a new component called GridPage, to
remove code duplication and defines two new subcomponents,
CollapsibleCardGrid and PaginatedCardGrid.

https://phabricator.endlessm.com/T31511