scripts: Copy all bundles from the template
The template now builds and bundles multiple zip files, so the
copy_bundles.py script is now adapted to copy all those zip files to
the corresponding destination. For example, an override named `foo`
will be bundled as `template-ui/foo.zip` and will be copied to:

```
DESTINATION/foo/custom-channel-ui.zip
```

Which is how the Explore plugin expects it.

Also: add description to the script to be displayed by calling --help.

https://phabricator.endlessm.com/T31531