template: Add link to carousel slides
This patch creates a new base component to use in all card kinds, the
BaseCard component is just a link to a node with an slot. This new
BaseCard is used in all different card kinds and also is used in the
Carousel component to have a link to the final content.

https://phabricator.endlessm.com/T31719