template: Rework the Card component composition
Replaced the BaseCard component with ContentLink that's a base link to
content with the goToContent methods there.

Now the cardMixin just has the thumbnail download logic and the card
title logic.

This patch also adds a new compoent (CarouselCard) to use this for the
Carousel component, right now each carousel card is just a
<b-carousel-slide> but in the future we can create more complex carousel
components and custom channels can customize just that part.

https://phabricator.endlessm.com/T31719