template: Custom CarouselCard style
This patch modifies the CarouselCard template to show the thumbnail to
the left and the description to the right.

This can't be done with a simple <b-card img-left> because that will
draw the image with it's size and not every thumbnails has the same
size, so to be able to adapt the thumbnail to the half of the card we
need to use the cover css property and the thumbnail as background.

https://phabricator.endlessm.com/T31719