template: Display topic title and description in header
Add a headerDescription getter. For the title we can use section.title
in all cases, which is the channel title for the tree root node.

Also remove SectionHeader from the Section view which shows the same
information. Keep the component for reference since it still has the
image background and the collapsible description.

https://phabricator.endlessm.com/T31767