template: Improve header icon and lead responsiveness
The `<b-img>` component expects the width in the template, not in the
CSS. So export the variable in the SCSS and import it in the
component, so we can use it in the template part.

Use the grid system for automatically adjusting the lead, instead of
using the CSS breakpoints.

Also, make the icon rounded as in latest design.

https://phabricator.endlessm.com/T31767