template: Add a new slidable card grid variant
And make it the default.

The cards are paginated into carousel slides of half size the default
pagination. A card group component is used, which makes all the cards
in the group the same tall.

The card group is a deck except for high quality card variant, which
have no gap space between.

Bootstrap card groups aren't responsive, so a FIXME is added to
consider smaller screens in upcoming commits.

Add custom buttons to the carousel, with same ARIA tags as the default
ones for accesibility.

Add invisible empty cards as needed to fill the last slide.

https://phabricator.endlessm.com/T31779