template: Add filters for structured tags
Add structured tags to the filter metadata.

Move the getter getStructuredTags logic to a function for reusing it.

Before the switch in possibleOptions, check if the filter name is
included in the structured tags. If so, sort and weight the values
obtained from the nodes.

And filter by structured tags if there is a query for any of them.

https://phabricator.endlessm.com/T31799