template: Navigate to content view
Move the logic from the App component to the store. Use a more
readable recursive function findNodeById to obtain the nodes.

Add content node to the navigation state.

Move tree from getter to state, because it's needed in mutations. The
tree is computed only once, when the channel information is set.

In App, use mapMutations instead of using $store.commit directly.

https://phabricator.endlessm.com/T31806