Template: Add unorm polyfill
The slugify library uses String.normalize, that's not supported on old
browsers like IE11 and there's no polyfill for this on babel/core-js.

This patch includes the unorm library, that implements the
String.normalize and could be used as polyfill for this feature to work
on IE11.

You can find more information here:
 * https://github.com/sindresorhus/slugify/issues/32
 * https://github.com/zloirock/core-js#missing-polyfills
 * https://stackoverflow.com/a/54242857

https://phabricator.endlessm.com/T31929