template: Use CardBody component on CarouselCard
I've tried to reuse the CardBody component, making configurable the
number of lines used by the title, so the card content is rendered by
the same component for all cards.

The css is done using a new sass function defined on styles.scss because
it's not easy to share vue component properties with css, the only way
is to use css variables and that's not working correctly with the math
operations done there and the css variables are not supported on IE11 so
we should not use that yet.

https://phabricator.endlessm.com/T31912